controller40 get/post 요청 핸들러 메소드2 작성

+++ src/main/java/com/study/domain/MyBean332.java
@@ -2,15 +2,14 @@ package com.study.domain;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Data
 public class MyBean332 {
-	private String stringCol;
-	private Integer intCol;
-	private BigDecimal decCol;
-	private LocalDate dateCol;
-	private LocalDateTime dateTimeCol;
+    private String stringCol; // VARCHAR
+    private Integer intCol; // INT
+    private Double decCol; // DEC
+    private LocalDate dateCol; // DATE
+    private LocalDateTime dateTimeCol; // DATETIME
 }
